change to array for tasks

+++ rand_t.h
@@ -23,7 +23,6 @@ struct rnd_t {
   auto make_seed_array() -> std::array< uint64_t, 5>
   {
     const auto e1 = static_cast<uint64_t>(std::chrono::high_resolution_clock::now().time_since_epoch().count());
-      // with some luck, this is non-deterministic: TRNG
     const auto e2 = static_cast<uint64_t>(std::random_device{}());
     // different between invocations from different threads within one app: thread-id
     const auto tid = std::this_thread::get_id();
@@ -31,7 +30,6 @@ struct rnd_t {
     // different between installations: compile time macros
     const auto e4 = static_cast<uint64_t>(std::hash<const char*>()(__DATE__ __TIME__ __FILE__));
     // likely different between runs, invocations and platforms: address of local
-    //  const auto e5 = reinterpret_cast<uint64_t>(&e1);
     return {{e1, e2, e3, e4, e4 ^ ~0ull /*e5*/}};
   }
 
@@ -50,23 +48,23 @@ struct rnd_t {
     return std::uniform_int_distribution<> (0, static_cast<int>(n - 1))(rndgen);
   }
 
-  double normal(double m, double s) {
-    std::normal_distribution<double> norm_dist(m, s);
+  float normal(double m, double s) {
+    std::normal_distribution<float> norm_dist(m, s);
     return norm_dist(rndgen);
   }
 
-  double threshold_normal() {
-    double output = threshold_dist(rndgen);
+  float threshold_normal() {
+    float output = threshold_dist(rndgen);
     while(output < 0) output = threshold_dist(rndgen);
     return output;
   }
 
   void set_threshold_dist(double m, double s) {
-    threshold_dist = std::normal_distribution<double>(m, s);
+    threshold_dist = std::normal_distribution<float>(m, s);
   }
 
 private:
-  std::normal_distribution<double> threshold_dist;
+  std::normal_distribution<float> threshold_dist;
 };
